requirements.txt
aiowsgi==0.7
anyio==3.3.2
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cloudpickle==1.6.0
contextlib2==21.6.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
gym==0.19.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
kiwisolver==1.3.2
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numba==0.54.0
numpy==1.20.3
numpy-quaternion==2021.8.30.10.33.11
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
pybullet==3.1.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySide2==5.15.2
PySimpleGUIQt==0.35.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.2.1
requests==2.26.0
requests-unixsocket==0.2.0
schema==0.7.4
scipy==1.7.1
Send2Trash==1.8.0
shiboken2==5.15.2
six==1.16.0
sniffio==1.2.0
stable-baselines3==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
waitress==2.0.0
wcfg==6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.2.1
zipp==3.5.0
